+++ src/components/Forms/SignUpForm/SignUpForm.js
@@ -1,0 +1,3 @@
+export const SignUpForm = () => {
+  return <div>dddd</div>;
+};
