+++ src/components/index.js
@@ -3,4 +3,5 @@ export * from "./Dropdown";
 export * from "./Input";
 export * from "./Label";
 export * from "./Overlay";
+export * from "./SignUpForm";
 export * from "./Textarea";
